requirements.txt
click  >= 7.0
numpy  >= 1.16.4
pandas >= 0.24.2
scipy  >= 1.3.0
pillow >= 6.1.0
matplotlib >= 3.1.1